Improve: add findItems function
This function is of the form:
`findItem("name", "type"[, exact[, case sensitive]])`
with the last two arguments defaulting to `true` if not given. Type
is one of "alias", "button", "keybind", "script", "timer" or "trigger".

This returns a list of all the IDs for the type of item whose name
matches the given "name". If the exact argument is `false` the given string
causes any item whose name has it as a substring to be returned in the
results. As a side effect if an empty string is supplied then all extant
items of the specified type will be returned - which might be useful in
some situations.

The need for this function was something I posited at the end of:
https://github.com/Mudlet/Mudlet/pull/6726#issue-1651092803

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>