Fix: get MSYS2/Mingw-w64 builds compiling
It seems that in an update a couple of `enum`s that we had put in for
Windows builds are NOT needed in this build environment and thus need to
be kept out of the build in this case.

This will be needed to test the set of three scripts I have been putting
together to setup, build and package Mudlet in that environment.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>