Improve: add persistent command history
This works for single word named command lines of all types, and saves them
all separately in the profile's base directory.

This should close #2007.

I am not yet happy with the save format - it does save the history in the
order that the entries are in in the
`(QStringList) TCommandLine::mHistoryList`
member in a UTF-8 form (which makes for easier external modification if
wanted (rather then Qt's default UTF-16 internal `QString`) but the
`QDataStream` format saves a `QByteArray` with a leading int32_t size
element - which, while good from a maintaining integrity point of view
makes it not amenable to external editing.

Testing also needs to be carried out in (possibly on a Window system) to
check that things still work with names that are more challenging for an
OS - e.g. including '\\' or '/' or ':'s...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>