Add: allow historical maps to be selected in settings (#6715)
#### Brief overview of PR changes/additions
This adds a `QComboBox` selector under the load map button in the mapper
tab on the "Profile preferences" dialogue to allow any map (Mudlet
`.dat` binary, `.json` or `.xml`) file stored in the "normal place" to
be chosen and loaded. Like the game save selection in the "Connect
profile" dialogue an icon is pre-pended to the "autosave" one, in
addition a (icon theme dependent) icon is pre-pended for `.json` and
`.xml` files so they can be distinguished from the "normal" binary one.
Like that game save list the entries are sorted into descending date and
time order.

#### Motivation for adding to Mudlet
It is a bounty issue.

#### Other info (issues closed, discussion etc)
This should close #656!

This PR originally ALSO added a similar `QComboBox` to the second tab of
the "Connect profile" dialogue - which I thought was a good idea but
which was declared to be outside of the scope of the original issue. I
will promulgate that option again in a future PR where we can discuss
how good an idea it really is! :smiling_imp:

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@hey.com>
Co-authored-by: mudlet-machine-account <39947211+mudlet-machine-account@users.noreply.github.com>