Revise: handle names for the command line history files better
We'll avoid using the names for the extra command lines as part of a
filename by instead storing them in a per profile `.ini` format file.

Using such a file to store per-profile is something that has been on my
TODO list for years (I prototyped code to do it many years ago) - so I have
added two functions to the Host class that hopefully can replace the
existing `readProfileData(...)` and `writeProfileData(...)`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>