Revise: add limit to number of command line history entries to save
Otherwise it will grow indefinitely large as every single entry is
retained.

A knob for this has been provided on the profile preferences and it is
saved with the profile's XML game save file (not in the base directory of
the profile. It covers the range of 0 to 500 entries with a default of 50
and a step size of 10. This value is applied to ALL command lines in a
profile.

Note that this limit is only applied when the commend history is saved,
it can still grow to be larger than the limit whilst the profile is active!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>