Revise: add per command-line controls for each of them
This adds three tool-buttons shown to the left of each command-line that,
in order:
* sets the command-line's most recently used commands (up to the
per-profile limit) to be saved in a file.
* sets the next command entered NOT to be saved in the history at all, as
a side effect of the way it works if the exact command has been used before
then entering it again with this option selected WILL erase that particular
entry from the stored history (and when the profile is closed) from the
file backing it up as well. After the command has been entered this option
is automatically deselected and the selected option reverts to the first
one.
* sets the command history for this command line to NOT be saved at all -
as Mudlet would behave before this PR came about. The commands are still
remembered for the current session but they are not written out to a file
though if they have been in a previous session than those will remain in
the file as it will not be updated when the profile is closed.

Also, if the number of lines to be saved is set to zero (the "None") value
in the settings then all three buttons are disabled (greyed out) in all the
command lines for that profile.

To add buttons to control this feature in the `QPlainTextEdit` proved to be
quite tricky. Unlike the `QLineEdit` class there is no provision to add
`QAction`s to either end. Instead I had to create a `TCommandLineWidget`
class to encapsulate a `TCommandLine` as well as the three buttons in a
`QFrame` container widget. Even this was not trivial but was possible to
do - unlike an initial failed attempt that tried to extend the
`TCommandLine` class itself to inherit from those other widgets as well.
The existing methods/interactions with the original `TCommandLine` had to
be revised to either work on the new wrapper class or to access the wrapped
one via that instead of working directly with it.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>