Infrastructure: migrate pull request creation to new github action (#6771)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Swap out gr2m/create-or-update-pull-request-action for a more popular
peter-evans/create-pull-request
#### Motivation for adding to Mudlet
We already used peter-evans/create-pull-request for creating the en-us
translations and it worked well. A far more popular action means it gets
more battle testing.
#### Other info (issues closed, discussion etc)
Minor change: instead of creating a unique branch per-commit, just one
branch is created now - this is so if we don't get a chance to merge a
PR for a week, it'll get updated with the latest content instead of a
separate, newer one being created.