Revise: remove the tool-buttons and move the UI to the context menu
This reduces the amount of disruption this PR was causing to the
`TCommandLine` class. I do have concerns about indicating to the end-user
that either the next or all commands are not being saved for the command
line going forward when we announce the feature that they will be saved
(by default) but I have not yet come up with a foolproof scheme to show
that yet.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>