Improve: Throw error when nonstring passed to f function (#6777)
#### Brief overview of PR changes/additions
Someone on one of the discord channels discovered the super secret variable when they passed a boolean in and it tried to :gsub it. This change does some type checking and throws an error if it's passed anything but a string.

#### Motivation for adding to Mudlet
Less ambiguous error than "attempt to index local 'supersecretstringvariablenocollision' (a boolean value)"

#### Other info (issues closed, discussion etc)
Also includes a test for the error.