Add deleteMultiline function to make deleting lines from a multiline trigger easier (#6779)
#### Brief overview of PR changes/additions
Adds the `deleteMultiline` function which will recursively delete all lines which have come in since the first trigger matched to start the multiline. This should not be used if there might be lines which came in since the trigger started matching but which you wish to keep.

#### Other info (issues closed, discussion etc)
Best way to use it is to pass the line delta of the multiline trigger it's called in. `deleteMultline(3)` if the line delta is 3, for instance.
Returns nil+error if called outside of the context of a multiline trigger. This felt better than halting script execution.