Fix: work around Discord user-name & discriminator changes
Discord is changing the way that Users are identified (to just lower case
alphanumeric characters and '-' and '.) and removing the *discriminator*.
However the unique user id (a long number) is not changing.

This PR encourages the user to migrate from using the first two to limit
which profile uses which Discord User-account to the latter. It causes
`[ ALERT ] ` messages to be shown on profile startup when the Discord-RPC
library is active and has made contact with the user's Discord application
and either a user specified user name or discriminator is being used to
limit. If they match up to the current Discord account in use the user id
number is shown and the user is requested to note it down and copy it to
the preferences and to then remove the previously entered name and
discriminator. Once this has been done (or if they have not been entered)
when a profile starts up a different message is given:
* an `[ INFO ] ` one that confirms that the Discord_RPC library is
connected and if a user id has been set nothing else, but if not the id is
shown so that it can be used in the future
* an `[ ALERT ] ` one that indicates that the entered id does not match the
one in use and what those numbers are

Once any existing preset user name or discriminator has been removed they
will be removed from the profile storage and the preference form bits for
them hidden and the text/tooltips for the visible id entry modified to
remove details about the older, "deprecated" details.

To enable the appropriate message to be give at the right time the way the
`(bool) Host::mIsProfileLoadingSequence` flag is accessed has been revised
so that it being reset triggers the message AND the `sysLoadEvent` Mudlet
event being raised, the latter is still done in the same point in the
application but it gets moved from the `mudlet` class to a called method in
the `host` class.

For the details of the Discord changes see their blog at:
https://discord.com/blog/usernames
and
https://support.discord.com/hc/en-us/articles/12620128861463

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>