add accessibility data for trigger list, say if active (#6778)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
For screen readers, there is not indication of if items in the trigger
editor are active or not. This change introduces that. Also some
inconsistent coloring of icons is improved.

#### Motivation for adding to Mudlet
closes #6166

#### Other info (issues closed, discussion etc)
I made some new functions to make Buttons, Scripts, and Timers use
similar image selection logic as Triggers and Aliases currently use.
That fixes the inconsistent graying of scripts disabled by parent and
that sort of thing. In a future PR it might be rearranged some so that
there aren't multiple spots doing the same thing, but for now I want
this to be easier to review, so I made it similar to existing setup. The
Variables list is already accessible so leaving that alone.

Discussion on Discord came up with phrasing "activated", "deactivated",
and adding the type of item in special cases. For example a trigger with
checkbox checked would be described as "activated" because you are
looking at page dedicated to triggers, while a trigger chain with the
funnel image would be "activated filter chain" because that is extra
information. Then if the trigger is under a parent that is turned off,
visualized as a checked box that is dimmed, it is called "X in a
deactivated group".

---------

Co-authored-by: atari2600tim <atari2600tim@users.noreply.github.com>