fix wrong description string selected for items with deactivated parents (#6850)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
In #6778 I used `descInactiveFolder` to describe (for screen readers) a
folder that is unchecked and
`descInactiveParent.arg(individualItemDescription)` for an item inside
of such a folder. So in a few places I have it saying description of a
folder itself, `deactivated folder`, instead of `%1 in a deactivated
group`, and then try to feed the individual description into a string
that has no spot for it. This should fix that.

#### Motivation for adding to Mudlet
Trying not to be evil I guess.

#### Other info (issues closed, discussion etc)
Credit to SlySven catching it