Fix: a wrong function name and a bogus error return value in setScript(...)
An error in #3515 meant that -1 and false would be returned as an error
message should the new lua code not compile instead -1 and an error
message.

The code was revised in #4536 as part of wider changes that made failure to
compile the lua script/code for all code setting functions throw a lua
error (aborting execution of the caller) with, what should be an error
message instead of a sentinel value (-1) and a message. However with the
prior error the error message will be "-1"!

In correcting the above I also discovered that the wanted error message
was: "permScript: cannot set script (error message)" for this function when
it should being with "setScript" and not "permScript"!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>