Improve: enabling and disabling all but main command line
Prevents crashing if `enableCommandLine(...)` or `disableCommandLine(...)`
are called with an empty string or "main" as the identifier.

Extends these two functions to show or hide (respectively) any other
`TCommandLine` instance whether created on the bottom of a `TConsole` (for
which the command line has the same name as the the console) with the first
of these functions or an extra one created on top of a `TConsole` and which
has been placed in an arbitrary position with the `createCommandLine(...)`
function.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>