Add overflow event for non scrolling consoles (#6872)
#### Brief overview of PR changes/additions
When a non-scrolling `TConsole` has text added to it that exceeds the
number of lines it can show this raises a `sysWindowOverflowEvent` with
the name of the console and the number of lines that have "scrolled off
the top" i.e. that have been hidden because there is not enough space
for them all. This can be used to manage the size of the text kept in
that window so that it is always fully visible.

#### Motivation for adding to Mudlet
This was prompted by the second part of my comment:
https://github.com/Mudlet/Mudlet/pull/6848#pullrequestreview-1425398702

#### Other issues
As I was looking at code that examines the number of lines of text in
the buffer I spotted that there were a couple of cases where the size of
the buffer or the size of a line in that buffer was being taken and then
being compared to be `<1` when really an `isEmpty()` test is a clearer
indication of intent - and where assignment to a temporary `(int)
length` variable could be made more streamlined.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@hey.com>
Co-authored-by: eowmob <72653851+eowmob@users.noreply.github.com>