improve error messages by making code start on first line (#6855)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Behind the scenes there are things like `code = qsl("function Alias%1()
%2\nend").arg(id, code)` which handles the code of an alias and embeds
it into a function. Timers, triggers, and keybinds all have similar, but
with the `function X()` on a line alone. When they get compiled, the
first line of an alias code box is also the first line of the resulting
function, so an error on the first line of the edit box code is also the
first line of the post-embed code. But for the other type of items, the
first line of edit box is second line of post-embed code. When
compiling, the system is looking at the post-embed code, so it gives
errors with line numbers that make sense with aliases but are confusing
with the others.

This edit should make them all similar to aliases.

#### Motivation for adding to Mudlet
fix #1460 
#### Other info (issues closed, discussion etc)
Error messages say things like "near X" instead of citing column
numbers, so I don't think there are any negative consequences for first
line being on same line after `function X()`. It's how aliases are
handled already.

---------

Co-authored-by: Stephen Lyons <slysven@virginmedia.com>