Improve: show script loading errors (#6868)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Added the ability to detect script loading errors, store them on the
script, and display them to the user in the script editor.

![Screenshot 2023-05-27 9 08 55
AM](https://github.com/Mudlet/Mudlet/assets/110988/7722654b-d7d2-436b-a47b-24d413d8398a)


#### Motivation for adding to Mudlet
This allows users to be notified if there were any issues loading their
scripts on profile load or after a profile reset. Without this, errors
would show just once as soon as a script was clicked on - and go away as
soon as the mouse was released. Displaying the errors in the UI allows
for the coder to find such errors easier - and hopefully know what to do
with them, since they are a common source of confusion in our #help
channel.

#### Other info (issues closed, discussion etc)
Please try the PR out in practice first before reviewing! Here's an
example to get started with:

```xml
<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE MudletPackage>
<MudletPackage version="1.001">
	<ScriptPackage>
		<ScriptGroup isActive="yes" isFolder="yes">
			<name>Referencing yet undeclared values test</name>
			<packageName></packageName>
			<script></script>
			<eventHandlerList />
			<Script isActive="yes" isFolder="no">
				<name>referencing undeclared value</name>
				<packageName></packageName>
				<script>a[1] = true</script>
				<eventHandlerList />
			</Script>
			<Script isActive="yes" isFolder="no">
				<name>declaring said value</name>
				<packageName></packageName>
				<script>a = {}</script>
				<eventHandlerList />
			</Script>
		</ScriptGroup>
	</ScriptPackage>
</MudletPackage>
```