fix: some UI messages that needs html escaping (#6859)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Fixed some more dynamic content that is displayed in a spot that handles
html but was not escaped for it. This batch is mostly file names and
folders and the blue warning about compile errors.

#### Motivation for adding to Mudlet
Progress toward #6795 

#### Other info (issues closed, discussion etc)
I pretty much only touched things that I could verify that I could put a
less than sign into. Someone else can probably figure out more.

Examples that are improved on:

![image](https://github.com/Mudlet/Mudlet/assets/29287358/2ae9af09-9a20-4f28-b3ae-b61727a3aa8c)

![image](https://github.com/Mudlet/Mudlet/assets/29287358/48b11a64-dc92-4f52-8fca-253b7ca9d76a)

![image](https://github.com/Mudlet/Mudlet/assets/29287358/cce34d33-94a9-4830-a935-e6fe64dab808)