Add persistent command history (#6767)
#### Brief overview of PR changes/additions
Saves each command line's history when a profile is closed and restores
the data the next time the profile is opened. Each command line is
handled separately.

#### Motivation for adding to Mudlet
It is a bounty item!

#### Other info (issues closed, discussion etc)
This works for command lines of all types, and saves them all separately
in the profile's base directory.

This should close https://github.com/Mudlet/Mudlet/issues/2007.

We'll avoid using the names for the extra command lines as part of a
filename by instead storing them in a `QSettings` per profile
`profile.ini` format file.

Using such a file to store per-profile details is something that has
been on my TODO list for years (I prototyped code to do it at least as
far back as 2016 I think) - so I have added two functions to the Host
class that hopefully can eventually replace the existing
`readProfileData(...)` and `writeProfileData(...)`.

Also added code to fix a corner case where a TCommandLine has a name
containing '/' or '\'.

Also added a limit to the number of command line history entries to
save. Otherwise it will grow indefinitely large as every single entry is
retained. A knob for this has been provided on the profile preferences
and it is saved with the profile's XML game save file (not in the base
directory of the profile). It covers the range of 0 to 10,000 entries
with a default of 500 and a logarithmic step size (1, 2, 5, 10) for each
multiple of 10. This value is applied to ALL command lines in a profile.

Note that this limit is only applied when the commend history is saved,
it can still grow to be larger than the limit whilst the profile is
active!

A pair of lua API  functions are added:
* `setSaveCommandHistory([commandLineName,] save)`  to enable
(`true`) or disable (`false`) the main, or if specified, any other
command-line
* `getSaveCommandHistory([commandLineName])` returns two values
first a boolean indicating whether the main, or if specified, any other
command-line will save its history (to a specified number set separately)
or not, between sessions and secondly a text message which either
reports that the history will be saved and how many entries or that it
will not and the reason why.

Also the `setConfig(...)`/`getConfig(...)` general settings functions gain
a new option `commandLineHistorySaveSize` which sets the number
of most recent commands stored for ALL command-lines in the given
profile for which saving is enabled .  The setter accepts a number of
values - which are duplicates of the values available on a new "knob"
in the Profile Preferences and the getter will report which of those
has been set. Setting the 'none' or '0' option will disable saving for
ALL command-lines so that they behave as they would before this
feature is introduced.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>