Infrastructure: setBackgroundColor adjusts a Label's stylesheet rather than palette (#6825)
#### Brief overview of PR changes/additions
Adjusts setBackgroundColor to use and adjust the stylesheet of a label
rather than using setPalette on the object. This should allow you to
adjust the background color of a label which uses a stylesheet.

Currently, once you set a stylesheet on a label setBackgroundColor stops
really doing anything. This way it should be a more consistent
experience.