Add: scrollingState() function to Lua API
The absence of a "getter" was noted by me in:
https://github.com/Mudlet/Mudlet/pull/6875#issuecomment-1572542949

This provides that function.

Note that this PR also makes the `(bool) TConsole::mScrollingEnabled` flag
private so that it cannot be manipulated by other classes except via the
accessor functions (including the one added herein) so that the additional
functionality in the setter (that clears any existing "split" - and sets
the `(bool) TTextEdit::mIsTailMode` flag for the upper pane in the console)
cannot be bypassed by those classes trying to modify the flag directory.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>