Fix: correct layout so that external translation comments get parsed
This replaces and closes #6886.

Whilst running `lupdate` to update `translations/translated/mudlet_en_US.ts`
it reported that some translation metadata could not be parsed for the
`[ XXXX ] - ` messages that `cTelnet::postMessage(...)` processes. Looking
more closely I could see that else if was being used but the external
comments were not in the correct location with the result that they
would not be correctly parsed by the `lupdate` tool. This commit fixes
them.

There was also a string marked for translation inside a lambda function in
the `Host` class that, because it was inside a lambda, was not able to
report the class it was in and thus was missing from the strings that
`lupdate` handled, by adding a class prefix this is now sorted out.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>