Add: clearMapSelection() function
This will close #6892 as it provides a function clearMapSelection() that
will clear and remove the selection of rooms (with the display of them in a
widget in the left of the 2D mapper widget). For functional reasons (that
it would be difficult to make work) it will fail and return a `nil` plus
and error message should the user be dragging the box to make a selection
at the time it is called. Otherwise it will return `true` or `false`
depending on whether a selection was cleared or not respectively.

When the code was put in place to pass the details of the selected rooms
back to the Lua sub-system via the `getMapSelection()` function it was
anticipated by me that that method might be extended to provide details
about other things that could be selected by the user in the 2D map.
However, up to now, this has not happened, but it might affect whether
providing this true or false return value is still feasible if it does.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>