Fix: opening profile automatically or from command line misses details
This should close #6753.

It was happening after #6709 because that stopped reading the Game Server
URL and port number from the Game Save data (as it should be read from the
per profile data stored in the profile's base directory and NOT, except for
special cases, the game save). However the details were being read and
stored in the `Host` instance in "normal" main case by the "Connect
Profile" dialogue but in the auto-load and for the command-line argument
cases that dialogue is not used and the details were not being read so the
URL was an empty string and the port was set to the default `23`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>