Improve: add separate tooltips to insertPopup() and echoPopup() functions
For these two functions the previous hint text was use both for the entries
in the right-click (context) menu AND the tool-tip that is shown when the
pointer is hovered over the link text. The corresponding `insertLink(...)`
and `echoLink(...)` functions only use the "hint" text for the hover tool-
tip and that can be written using Qt's Rich-Text (in the same manner as
"Labels" can) to be quite fancy. However without this PR doing the same
for the "hints" for these two functions means all the formatting code would
be used for the context menu entries as well. With this PR those tool-tips
can (optionally, so that existing scripts behave the same) be provided in a
separate table (with the same number of entries).

If the command (string or function form) and the hint entries are empty
strings `""` then a context menu entry is not made for that one but it can
still have a tooltip - which is a means to introduce content for the
tool-tip independent of any particular item. Furthermore if the command
(string or function form) is an empty string (so there is nothing to do)
but there **is** a hint then that hint text is entered into the context menu
as a disabled (inactive) item - which might be useful in some situations.

Previously when the popup (context, right click, menu) is generated and
displayed a temporary `QMap` is created to hold the commands or functions
that all the `QAction`s will need to call if activated. It did use the
"hint" or context menu text for each item as the key - however this would
not handle duplicate texts. I have switched to storing the entry number
of the items in the tables supplied (starting at `1` not `0`) as the key
for that table instead and store the same value as the `Qt::UserRole` data
item for each `QAction` - this will now work should the same text be used
for more than one "hint".

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>